batch_job/onprem_batch_job/requirements.txt
psycopg2-binary
python-decouple==3.8
pandas==2.0.3
pyarrow==13.0.0
fsspec==2023.6.0
gcsfs==2023.6.0
google-cloud-storage